Pull event registry from fsm state into its own file
