+++ Assets/Code/_Common/EventRegistry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3f3ccf44349a8a42ba33af711513a41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
