+++ Assets/Code/_Common/FsmState.cs
@@ -15,45 +15,6 @@ namespace PQ.Common
     */
     public abstract class FsmState : IEquatable<FsmState>
     {
-        // todo: look into extracting this out into some sort of bus, and replace GameEventCenter with it
-        private class EventRegistry<EventData>
-        {
-            private string _description;
-            private readonly Dictionary<GameEvent<EventData>, Action<EventData>> _eventToHandlerMapping;
-
-            public override string ToString() => _description;
-
-
-            public EventRegistry(params (GameEvent<EventData>, Action<EventData>)[] eventCallbacks)
-            {
-                var stringBuilder = new StringBuilder(eventCallbacks.Length);
-                _eventToHandlerMapping = new Dictionary<GameEvent<EventData>, Action<EventData>>(eventCallbacks.Length);
-                foreach (var (event_, callback_) in eventCallbacks)
-                {
-                    _eventToHandlerMapping[event_] = callback_;
-                    stringBuilder.AppendFormat("{0}=>{1};", event_.Name, callback_.Method.Name);
-                }
-
-                _description = stringBuilder.ToString();
-            }
-
-            public void StartListening()
-            {
-                foreach (var (event_, callback_) in _eventToHandlerMapping)
-                {
-                    event_.AddListener(callback_);
-                }
-            }
-            public void StopListening()
-            {
-                foreach (var (event_, callback_) in _eventToHandlerMapping)
-                {
-                    event_.RemoveListener(callback_);
-                }
-            }
-        }
-
-
         private readonly string _name;
         private bool _isActive;
         private bool _isHookedupWithEvents;
